Use the new parser for AsMut as well
